+++ src/intel/compiler/ibc_live_intervals.c
@@ -48,7 +48,7 @@ reg_ref_has_live_data(const ibc_reg_ref *ref, const ibc_live_intervals *live)
 
 static bool
 record_reg_write_sizes(ibc_reg_ref *ref,
-                       UNUSED int8_t num_bytes, UNUSED int8_t num_comps,
+                       UNUSED int num_bytes, UNUSED int num_comps,
                        UNUSED uint8_t simd_group, uint8_t simd_width,
                        void *_state)
 {
@@ -119,7 +119,7 @@ reg_num_chunks(const ibc_reg *reg, ibc_live_intervals *live)
 void
 ibc_live_intervals_reg_ref_chunks(const ibc_live_intervals *live,
                                   const ibc_reg_ref *ref,
-                                  int8_t num_bytes, int8_t num_comps,
+                                  int num_bytes, int num_comps,
                                   uint8_t simd_group, uint8_t simd_width,
                                   BITSET_WORD *chunks)
 {
@@ -362,7 +362,7 @@ struct setup_use_def_state {
 
 static bool
 setup_block_use_def_for_read(ibc_reg_ref *ref,
-                             int8_t num_bytes, int8_t num_comps,
+                             int num_bytes, int num_comps,
                              uint8_t simd_group, uint8_t simd_width,
                              void *_state)
 {
@@ -393,7 +393,7 @@ setup_block_use_def_for_read(ibc_reg_ref *ref,
 
 static bool
 setup_block_use_def_for_write(ibc_reg_ref *ref,
-                              int8_t num_bytes, int8_t num_comps,
+                              int num_bytes, int num_comps,
                               uint8_t simd_group, uint8_t simd_width,
                               void *_state)
 {
@@ -518,7 +518,7 @@ struct extend_live_interval_state {
 
 static bool
 extend_live_interval_for_read(ibc_reg_ref *ref,
-                              int8_t num_bytes, int8_t num_comps,
+                              int num_bytes, int num_comps,
                               uint8_t simd_group, uint8_t simd_width,
                               void *_state)
 {
@@ -553,7 +553,7 @@ extend_live_interval_for_read(ibc_reg_ref *ref,
 
 static bool
 extend_live_interval_for_write(ibc_reg_ref *ref,
-                               int8_t num_bytes, int8_t num_comps,
+                               int num_bytes, int num_comps,
                                uint8_t simd_group, uint8_t simd_width,
                                void *_state)
 {
